requirements.txt
numpy
starred-astro
lightcurver
astropy
scipy
pyvo
lsst-daf-butler
lsst-sphgeom


